refactor Windows 10 fix to fixWindows10GulpPathIssue and implement it for copy  and util
