+++ gulp/util/util.js
@@ -1,4 +1,5 @@
 import fs from 'fs';
+import path from 'path';
 // foldero - will load files from the folder you specify and construct an object
 // with properties corresponding to each loaded file
 import foldero from 'foldero';
@@ -40,13 +41,44 @@ const getImageCollection = obj => {
       obj.plugins.imagemin.svgo({plugins: [{removeViewBox: true}]})
     ])
   ))
+  // .pipe(plugins.debug())
+
+  // Fix for Windows 10 and gulp acting crazy
+  .pipe(obj.plugins.rename(file => {
+    const {dest, plugins, config} = obj;
+    fixWindows10GulpPathIssue({file, dest, plugins, config})
+  }))
+  // .pipe(plugins.debug())
   .pipe(obj.gulp.dest(obj.dest));
 };
 
-const getStaticFiles = ({staticFilePath, dest, gulp, plugins}) => {
+const fixWindows10GulpPathIssue = ({file, dest, plugins, config}) => {
+  let dirList = file.dirname.split(path.sep);
+  let destList = dest.split(path.sep);
+  if (dirList.length === 1) {
+    file.dirname = '';
+  }
+
+  if (file.dirname.indexOf('\\') !== -1) {
+    if (dirList[0] === config.directory.source) {
+      dirList.shift();
+    }
+
+    dirList = dirList.map(x => x.replace(/\_/, '')).filter(x => !destList.includes(x));
+    file.dirname = dirList.join(path.sep);
+  }
+};
+
+const getStaticFiles = ({staticFilePath, dest, gulp, plugins, config}) => {
   return gulp
     .src(staticFilePath)
     .pipe(plugins.changed(dest))
+
+    // Fix for Windows 10 and gulp acting crazy
+    .pipe(plugins.rename(file => {
+      fixWindows10GulpPathIssue({file, dest, plugins, config})
+    }))
+
     .pipe(gulp.dest(dest));
 };
 
@@ -54,5 +86,6 @@ export {
   getStaticFiles,
   getJsonData,
   getImageCollection,
-  printError
+  printError,
+  fixWindows10GulpPathIssue
 };
